Release v1.0.5

+++ fitness_app/lib/modules/main/modules/home/modules/program/detail/widgets/program_overview.dart
@@ -52,7 +52,7 @@ class ProgramOverview extends StatelessWidget {
             Expanded(
               child: ProgramInfoTile(
                 icon: Ionicons.fitness,
-                label: totalCalo.toString(),
+                label: totalCalo.toInt().toString(),
                 iconColor: AppColors.error,
                 backgroundColor: AppColors.errorSoft,
               ),
